Move collective functions to data_reader
