+++ src/fairseq2/datasets/_data_reader.py
@@ -10,15 +10,16 @@ from abc import ABC, abstractmethod
 from collections.abc import Iterator, Mapping
 from typing import TypeVar, final
 
+import torch
 from typing_extensions import Self, override
 
 from fairseq2.data import DataPipeline, DataPipelineError
-from fairseq2.gang import Gang, GangError
+from fairseq2.gang import Gang, GangError, all_sum
+from fairseq2.logging import log
 
 # isort: split
 
 from fairseq2.datasets._config import DataReadOptions, SyncMode
-from fairseq2.datasets._utils import _min_num_batches, _sum_num_batches
 
 BatchT_co = TypeVar("BatchT_co", covariant=True)
 
@@ -96,7 +97,9 @@ class DataPipelineReader(DataReader[BatchT]):
         :param pipeline: The data pipeline to iterate over.
         :param gang: The gang over which the underlying dataset is sharded.
         :param options: The read options.
-        :param strict_state: If ``True``, the state of the data pipeline is strict (taking state from all pipeline steps).
+        :param strict_state: If ``True``, the entire state of the data pipeline
+            including shuffling and bucketing buffers will be included in the
+            state dictionary.
         """
         self._dataset_name = dataset_name
         self._split = split
@@ -192,3 +195,32 @@ class DataPipelineReader(DataReader[BatchT]):
     @override
     def num_accumulate(self) -> int:
         return self._options.num_accumulate
+
+
+def _min_num_batches(num_batches: int, gang: Gang) -> int:
+    all_num_batches = torch.zeros((gang.size,), device=gang.device, dtype=torch.int64)
+
+    input_ = torch.tensor([num_batches], device=gang.device)
+
+    gang.all_gather(all_num_batches, input_)
+
+    min_num_batches = int(all_num_batches.min())
+    if min_num_batches != 0:
+        return min_num_batches
+
+    # If not all processes have reached end of data, report the ones that have
+    # reached for debugging purposes.
+    if log.is_enabled_for_debug() and all_num_batches.sum() > 0:
+        ranks = all_num_batches.bool().logical_not_().nonzero().squeeze(-1).tolist()
+
+        s = ", ".join(str(r) for r in ranks)
+
+        log.debug("End of data reached at rank(s) {}.", s)
+
+    return 0
+
+
+def _sum_num_batches(num_batches: int, gang: Gang) -> int:
+    total_num_batches = all_sum(gang, num_batches)
+
+    return int(total_num_batches)
